requirements.txt
streamlit
pandas
plotly
eurostat
requests
python-dotenv
entsoe-py